#7 Add a custom exception

+++ src/main/java/movie/core/repository/ActorRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ActorRepository extends CrudRepository<Actor, Integer> {
+public interface ActorRepository extends CrudRepository<Actor, Long> {
 
 }
